#11015 create Login/ResetPassword routes components - remove navigation middleware
